+++ apps/api/views/branch_views.py
@@ -54,7 +54,7 @@ def read_branch_list_view(request, pk):
         company = get_company_by_id_service(user, pk)
         if company:
             serializer = BranchSerializer(company.branchs, many=True)
-            message = 'Запрос прошел успешно'
+            message = 'Запрос выполнен успешно'
             logger.info(f'{__name__} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else:
@@ -109,7 +109,7 @@ def read_branch_view(request, pk):
         branch = get_branch_by_id_service(user=user, branch_id=pk)
         if branch:
             serializer = BranchSerializer(branch, many=False)
-            message = 'Запрос прошел успешно'
+            message = 'Запрос выполнен успешно'
             logger.info(f'{__name__} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else:
@@ -223,7 +223,7 @@ def create_branch_view(request):
             serializer = BranchSerializer(branch, data=request.data, partial=True)
             serializer.is_valid()
             serializer.save()
-            message = 'Запрос прошел успешно'
+            message = 'Запрос выполнен успешно'
             logger.info(f'{__name__} - {message} / user id:{user.id}')
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
@@ -334,7 +334,7 @@ def update_branch_view(request, pk):
             serializer = BranchSerializer(branch, many=False, partial=True, data=data)
             if serializer.is_valid():
                 serializer.save()
-            message = 'Запрос прошел успешно'
+            message = 'Запрос выполнен успешно'
             logger.info(f'{__name__} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else: